fix(userRanking): fix the return value of the ratio function to not 'floor' the divisiongit status

+++ app/src/main/java/com/github/se/orator/model/profile/UserProfileViewModel.kt
@@ -650,7 +650,7 @@ class UserProfileViewModel(internal val repository: UserProfileRepository) : Vie
           totalNbrSessions != null &&
           nbrSuccess >= 0 &&
           totalNbrSessions > 0) {
-        return (nbrSuccess / totalNbrSessions).toDouble()
+        return nbrSuccess.toDouble() / totalNbrSessions.toDouble()
       }
       return -1.0
     } else {
